bomber

+++ src/Position.java
@@ -1,29 +1,29 @@
 public class Position implements Cloneable {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     Position() {
 
     }
 
-    Position(int x, int y) {
+    Position(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
